(#5272) fix checkpointing against Cloudant/CouchDB 2.0
An attempt to provide a better fix for #5145. It seems like
it's safe to eagerly write a checkpoint in the event there are no
changes returned (e.g. because a filter did not generate results).

This at least allows the tests to pass - possibly this means we
need more tests.